Add "clear:right" to right-floating thumbnails, and "clear:left" to left-floating ones.
This solves a problem we've had for some time: Two right-floating images with too little text
between them cascade inwards into the middle of the page. Adding more text isn't a viable
solution because "too little text" is more and more text with increasing screen resolution. This
solves it elegantly, I think. Also, since the 'clear' is added only to the right-floating
elements, this does *not* push the text down and does *not* cause blocks of whitespace to occur.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@4112 dd0e9695-b195-4be7-bd10-2dea1a65a6b6