* (bug 14515) HTML nesting cleanup on edit form
The edit form was being a bit wonky about how it nested certain things, splitting the "summary" *label* from the input form and following check boxes with a <div> boundary... which was then forced to be inline.

This maybe made sense long ago, but seems pretty silly now. The div boundary made even less sense when in 'add comment' mode, and when hiding #edit-form via custom JS or css it would abut the <textarea> directly against the following checkboxes, leading to intermittent layout problems.

Moved the <div> to surround the summary box (for regular editing) and down below the textarea (for add-comment mode) so editOptions makes some kind of sense... hopefully....



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@36515 dd0e9695-b195-4be7-bd10-2dea1a65a6b6