Add non-identity collation, with migration script
It seemed to work correctly, with the newly-created page "bob" sorting
as "BOB", but then I nuked all my cl_sortkey by running the migration
script before refreshLinks.php had finished running, so I'll have to
wait a while to see if it works properly with a non-messed-up database.
It's possible there's something wrong with the display of section
letters in the categories, but otherwise I think this is working right.

git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@69816 dd0e9695-b195-4be7-bd10-2dea1a65a6b6