Dual strategy thumbnailing -- locally for development and simpler wikis, or in the cluster for setups like the WMF's
When we did our test deploy, we found that we could not scale thumbnails locally in the cluster (and this was undesirable anyway).
So, we moved UploadStash thumbnailing to occur on URL invocation, which is more like the usual MediaWiki model anyway. So the customized transform()
moved from UploadStash to just being a special case of how SpecialUploadStash does things.

Note that the Special:UploadStash url masks how the thumbnail is obtained. Unlike the regular Commons wiki, the user never sees the cluster's URL for the thumbnail.
A web request, or an imageMagick call, is performed right there and then, and the results are catted out to the client.

For consistency, we did not use wfStreamfile since we were in some cases streaming from contents obtained over the MWhttpRequest, not just local files.



git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@77451 dd0e9695-b195-4be7-bd10-2dea1a65a6b6