(follow-up r80554) If person typed max amount in edit summary, backspace didn't work on some versions of opera
This still doesn't recognize all special keys on affected versions of opera as
keystrokes to ignore. I don't think this is a big issue, since backspace
is recognized (which is the only key thats really important), newer
versions of Opera should not be affected, and (based on a skim of the jQuery
site) it appears that if we ever upgrade our version of jQuery, the issue
might go away. (not 100% sure about that last part)

Originally when i was testing this code, I tested in all sorts of browsers
that it would stop people from putting more text in. I never tested that
it would allow backspaces (I now have, opera was the only affected one where
it didn't work).

Thanks to Alex Smotrov for reporting issue.


git-svn-id: http://svn.wikimedia.org/svnroot/mediawiki/trunk@83067 dd0e9695-b195-4be7-bd10-2dea1a65a6b6